+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 )
 
 ## 0.0.4
-- Added _Content Security Policy_, (Issue #3), and associated `quickbook.preview.contentSecurityPolicy` setting.
+- Added _Content Security Policy_, (Issue #3), and associated `quickbook.preview.security.contentSecurityPolicy` setting.
 - Modified settings functionality to reload with every _preview_ operation (no more _Reload Window_ necessary).
 - Partially fixed Issue #2, with support for `CSS` file setting & support graphics (not user-images).
     > Note that at the time of writing I needed to set the _Graphics Path_ setting to: